Create GRAPH
Students become friends of each other in the school and form groups. Friendship between
Student X ​and Student Y​ is represented as C X Y​. When two students belonging to different
group become friends, the net effect is the merger of both groups which X​ and Y​ belongs to. At
the beginning, there are N​ students representing N​ groups. Suppose student 1 and 2 become
friends and later 2 and 3 become friends, then 1,2, and 3 will belong to the same group.
There are two type of queries:
1. C I J => groups containing student I and J merged (if they belong to different groups).
2. Q I => print the size of the group to which student I​ belongs.